Shalev's_Projects/5_StickerBot/requirements.txt
opencv-python
rembg
onnxruntime
Pillow
